Added logic to the game
